Sanity testing
The text "About your company" does not have a two-sided equality. The equality is given only on the left side. It is also advisable to review the ending of the text, as it ends with a comma. It would be more beautiful if it were shortened and ended with a period.